test/requirements_lock.txt
plumbum==1.7.2
PyYAML==6.0
